+++ src/firmware/app/world/firmware_robot_test.cpp
@@ -7,7 +7,7 @@ extern "C"
 
 #include <gtest/gtest.h>
 
-#include "software/test_util/test_util.h"
+#include "shared/test_util/test_util.h"
 
 class FirmwareRobotTest : public testing::Test
 {
