+++ src/software/test_util/test_util.cpp
@@ -110,15 +110,4 @@ namespace TestUtil
         }
         return states;
     }
-
-    void handleTestRobotLog(TbotsProto_RobotLog robot_log)
-    {
-        LOG(INFO) << "[TEST ROBOT " << robot_log.robot_id << "]["
-                  << TbotsProto::LogLevel_Name(
-                         static_cast<TbotsProto::LogLevel>(robot_log.log_level))
-                  << "]"
-                  << "[" << robot_log.file_name << ":" << robot_log.line_number
-                  << "]: " << robot_log.log_msg;
-    }
-
 };  // namespace TestUtil
