+++ src/software/test_util/test_util.h
@@ -129,12 +129,4 @@ namespace TestUtil
      */
     std::vector<RobotStateWithId> createStationaryRobotStatesWithId(
         const std::vector<Point> &positions);
-
-    /**
-     * Takes a RobotLog proto and creates an INFO level g3log
-     *
-     * @param robot_log The RobotLog to log
-     */
-    void handleTestRobotLog(TbotsProto_RobotLog robot_log);
-
 };  // namespace TestUtil
